data_aggregation/requirements.txt
boto3==1.38.13
botocore==1.38.13
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
urllib3==1.26.20
